# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WorkNest Backend API",
    "version": "1.0.0",
    "description": "Canonical HTTP contract for the WorkNest backend service."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/v1/jobs": {
      "post": {
        "tags": ["Jobs"],
        "summary": "Create job",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "allOf": [
                  { "$ref": "#/components/schemas/JobCreateRequest" },
                  {
                    "type": "object",
                    "properties": {
                      "companyLogo": { "type": "string", "format": "binary" }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": { "201": { "$ref": "#/components/responses/Success" } }
      }
    },
...
    "/api/v1/applications": {
      "get": {
        "tags": ["Applications"],
        "summary": "Get all applications (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/ApplicationStatus" }
          },
          {
            "name": "jobId",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/ObjectId" }
          },
          {
            "name": "applicantId",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/ObjectId" }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" }
          },
          { "$ref": "#/components/parameters/KeywordQuery" },
          { "$ref": "#/components/parameters/PageQuery" },
          { "$ref": "#/components/parameters/AppLimitQuery" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "tags": ["Notifications"],
        "summary": "List notifications",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/PageQuery" },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
          },
          { "name": "isRead", "in": "query", "schema": { "type": "boolean" } },
          { "name": "unreadOnly", "in": "query", "schema": { "type": "boolean" } }
        ],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken"
      }
    },
    "parameters": {
...
      "KeywordQuery": {
        "name": "keyword",
        "in": "query",
        "schema": { "type": "string", "maxLength": 120 }
      },
...
      "PageQuery": {
        "name": "page",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "default": 1 }
      },
...
      "AppLimitQuery": {
        "name": "limit",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 10 }
      }
    },
    "responses": {
      "Success": {
        "description": "Successful response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiResponse" }
          }
        }
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "message": { "type": ["string", "null"] },
          "data": {}
        },
        "required": ["success", "message", "data"]
      },
      "ObjectId": {
        "type": "string",
        "pattern": "^[0-9a-fA-F]{24}$"
      },
...
      "JobType": {
        "type": "string",
        "enum": ["Full-Time", "Contract", "Part-Time", "Internship", "Freelance"]
      },
      "JobStatus": {
        "type": "string",
        "enum": ["active", "draft", "closed"]
      },
      "ApplicationStatus": {
        "type": "string",
        "enum": ["submitted", "in_review", "shortlisted", "interview", "offer", "rejected", "hired"]
      },
...
      "SalaryRange": {
        "type": "object",
        "properties": {
          "min": { "type": "number", "minimum": 0 },
          "max": { "type": "number", "minimum": 0 }
        },
        "required": ["min", "max"]
      },
      "JobBase": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "minLength": 3, "maxLength": 120 },
          "location": { "type": "string", "minLength": 2, "maxLength": 120 },
          "jobType": { "$ref": "#/components/schemas/JobType" },
          "category": { "type": "string", "minLength": 2, "maxLength": 80 },
          "experienceLevel": { "type": "string", "minLength": 2, "maxLength": 60 },
          "jobDescription": { "type": "string", "minLength": 20, "maxLength": 5000 },
          "responsibilities": {
            "type": "array",
            "items": { "type": "string", "minLength": 2, "maxLength": 500 },
            "minItems": 1
          },
          "requirement": {
            "type": "array",
            "items": { "type": "string", "minLength": 2, "maxLength": 500 },
            "minItems": 1
          },
          "benefits": {
            "type": "array",
            "items": { "type": "string", "maxLength": 500 },
            "default": []
          },
          "salaryRange": { "$ref": "#/components/schemas/SalaryRange" },
          "companyName": { "type": "string", "minLength": 2, "maxLength": 120 },
          "companyWebsite": { "type": "string", "format": "uri" },
          "applicationQuestions": {
            "type": "array",
            "items": { "type": "string", "maxLength": 300 },
            "default": []
          },
          "status": { "$ref": "#/components/schemas/JobStatus" }
        }
      },
      "JobCreateRequest": {
        "allOf": [
          { "$ref": "#/components/schemas/JobBase" },
          {
            "type": "object",
            "required": [
              "title",
              "location",
              "jobType",
              "category",
              "experienceLevel",
              "jobDescription",
              "responsibilities",
              "requirement",
              "salaryRange",
              "companyName"
            ],
            "properties": {
              "status": {
                "type": "string",
                "enum": ["active", "draft", "closed"],
                "default": "draft"
              }
            }
          }
        ]
      },
...
    }
  }
}